Load configs correctly, refactor for better control and results
Using appsettings.json for configs, loading and validating it to make it simpler to use. Refactored classes to use better dependency injection for testability.

Attempted to compress index so that multiple long paths are hashed and serialized.

Added results processor so ids with multiple token matches get better priority and output constructed easily.