Update InvertedIndex.cs to new version with smarter compression logic, updated indexing to remove special characters and stop words, clean up code
When building an index we now determine whether compression is needed dynamically by checking repeated id count vs non-repeated. This will help avoid building a mapping when not needed for smaller indexes.

When building index tokens kept are now being cleaned to remove characters such as .'"! as they interfere with good results. Also, added stop words to the config with some basic ones loaded at start up so that useless words are not stored in the index.

Cleaned up code with some help text, added argument validation and removal of unused of code.