Include filenames in indexing and cleaned repeated code
When indexing a dir we should also index the filenames of files with/without a valid extension.
Updated the function to get valid filepaths to instead return path to all files with a named tuple that uses a
flag to distinguish between filepaths to index content and those we just need the filename.

Also cleaned up repeated code in FileUtilities.cs where file is checked for valid extension. In InvertedIndex.cs the
 indexing of a line of text was extracted to a method so we don't repeated code between content and filename indexing.