Add checksum for index contents+filepaths, include readme
When indexing files we will now also generate a MD5 hash of file contents + the file path
and combine it will the prev hash checksum value (from the previous) file to get a checksum
o detect changes in either contents of files used in the index or the file names.

Readme added with information about the project and media for easier understanding.