Improve set compression logic
The logic so decide whether to compress the index in memory and for storage was
quite basic and relied only repeated vs non-repeated ids/paths. This is not
accurate as what we want to achieve is reduced size. To achieve this a better comparision
is made between the size of the compressed ids using long numbers vs the text length of all
repeated ids as-is.