Use GZip for serializing index
First version index stored was 100kb. Using a id mapping table to compress data where the mapping used
was the hash of the id(path) took the index to 161kb. Then, using number instead of hashes reduced the
index from 160kb to 86kb. Further in-memory/storage will need to be considered but for reducing storage
space further using gzip when writing the index reduced it to 25kb.